+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Adjust checkout columns width. Fixes UICHKOUT-457.
 * Add items awaiting pickup modal. Part of UICHKOUT-475.
-
+* Display modal for multipiece items. Part of UICHKOUT-446.
 
 ## [1.4.0](https://github.com/folio-org/ui-checkout/tree/v1.4.0) (2018-12-13)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v1.3.1...v1.4.0)